refactor: move ClientInfo out
