+++ docs/client_info.md
@@ -1,0 +1,3 @@
+# ClientInfo
+
+::: client_info.ClientInfo
