+++ src/soccerxcomm/__init__.py
@@ -1,8 +1,8 @@
 from .client import Client
+from .client_info import ClientInfo
 from .game_stage_kind import GameStageKind
 from .http_client import HttpClient
 from .http_server import HttpServer
-from .logger import Logger
 from .message import Message
 from .network_client import INetworkClient
 from .network_server import INetworkServer
@@ -10,6 +10,7 @@ from .server import Server
 
 __all__ = [
     "Client",
+    "ClientInfo",
     "GameStageKind",
     "HttpClient",
     "HttpServer",
